backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
PyPDF2==3.0.1
python-docx==1.1.0
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.3
joblib==1.3.2
Werkzeug==2.3.7
